+++ lib/rdf/serializations/ntriples.ex
@@ -20,6 +20,7 @@ defmodule RDF.NTriples do
   import RDF.Sigils
 
   @id           ~I<http://www.w3.org/ns/formats/N-Triples>
+  @name         :ntriples
   @extension    "nt"
   @content_type "application/n-triples"