+++ lib/rdf/serializations/turtle.ex
@@ -11,6 +11,7 @@ defmodule RDF.Turtle do
   import RDF.Sigils
 
   @id           ~I<http://www.w3.org/ns/formats/Turtle>
+  @name         :turtle
   @extension    "ttl"
   @content_type "text/turtle"